Rename containers to pages
